# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_tumbuhindigi-sys_momiji-home-backend_internal_cart.CartItem:
    properties:
      balance_due:
        type: string
      deposit_amount:
        type: string
      id:
        type: string
      image_src:
        type: string
      inventory_quantity:
        type: integer
      quantity:
        type: integer
      subtotal:
        type: string
      title:
        type: string
      unit_price:
        type: string
      variant_id:
        type: string
    type: object
  github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.ErrorBlock'
      message:
        type: string
      status:
        description: '"success" or "error"'
        type: string
      timestamp:
        type: string
    type: object
  github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.ErrorBlock:
    properties:
      code:
        type: string
      details: {}
    type: object
...
  internal_cart.CartItem:
    properties:
      balance_due:
        type: string
      deposit_amount:
        type: string
      id:
        type: string
      image_src:
        type: string
      inventory_quantity:
        type: integer
      quantity:
        type: integer
      subtotal:
        type: string
      title:
        type: string
      unit_price:
        type: string
      variant_id:
        type: string
    type: object
  internal_cart.CartItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
      variant_id:
        type: string
    required:
    - quantity
    - variant_id
    type: object
  internal_cart.CartResponse:
    properties:
      pre_order:
        items:
          $ref: '#/definitions/internal_cart.CartItem'
        type: array
      session_id:
        type: string
      ship_ready:
        items:
          $ref: '#/definitions/internal_cart.CartItem'
        type: array
      summary:
        $ref: '#/definitions/internal_cart.CartSummaryDTO'
    type: object
  internal_cart.CartSummaryDTO:
    properties:
      currency:
        type: string
      total_balance_due:
        type: string
      total_charged_now:
        type: string
      total_deposit:
        type: string
      total_pre_order:
        type: string
      total_ship_ready:
        type: string
    type: object
...
  internal_cart.MergeCartRequest:
    properties:
      guest_session_id:
        type: string
    required:
    - guest_session_id
    type: object
  internal_cart.UpdateCartItemRequest:
    properties:
      quantity:
        minimum: 0
        type: integer
    type: object
...
  internal_checkout.CheckoutSummaryRequest:
    properties:
      address_id:
        type: integer
      shipping_method:
        type: string
    required:
    - shipping_method
    type: object
  internal_checkout.CheckoutSummaryResponse:
    properties:
      currency:
        type: string
      due_august:
        properties:
          preorder_balance:
            type: string
          shipping_preorder:
            type: string
          total:
            type: string
        type: object
      due_now:
        properties:
          preorder_deposit:
            type: string
          ship_ready_total:
            type: string
          shipping:
            type: string
          total:
            type: string
        type: object
      pre_order:
        properties:
          balance_subtotal:
            type: string
          deposit_subtotal:
            type: string
          items:
            items:
              $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_cart.CartItem'
            type: array
        type: object
      ship_ready:
        properties:
          items:
            items:
              $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_cart.CartItem'
            type: array
          subtotal:
            type: string
        type: object
      shipping:
        properties:
          cost:
            type: string
          estimated_arrival:
            type: string
          method:
            type: string
        type: object
    type: object
  internal_checkout.InitiateCheckoutRequest:
    properties:
      address_id:
        type: integer
      address1:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      shipping_method:
        type: string
      state:
        type: string
      zip:
        type: string
    required:
    - shipping_method
    type: object
  internal_checkout.InitiateCheckoutResponse:
    properties:
      checkout_url:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Momiji Home Backend API with Shopify orchestration
  termsOfService: http://swagger.io/terms/
  title: Momiji Home API
  version: "1.0"
paths:
...
  /cart:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Clear all cart items
      tags:
      - Cart
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_cart.CartResponse'
              type: object
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Get Cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Add Item Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cart.CartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/items/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    patch:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Quantity Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cart.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Update item quantity
      tags:
      - Cart
  /cart/merge:
    post:
      consumes:
      - application/json
      parameters:
      - description: Merge Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cart.MergeCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Merge guest cart into user cart
      tags:
      - Cart
...
  /cart/summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_cart.CartSummaryDTO'
              type: object
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Get Cart Summary
      tags:
      - Cart
  /checkout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Initiate Checkout Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_checkout.InitiateCheckoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_checkout.InitiateCheckoutResponse'
              type: object
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Initiate Shopify checkout
      tags:
      - Checkout
  /checkout/confirm:
    get:
      parameters:
      - description: Shopify Order ID
        in: query
        name: shopify_order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get checkout confirmation status
      tags:
      - Checkout
  /checkout/summary:
    post:
      consumes:
      - application/json
      parameters:
      - description: Checkout Summary Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_checkout.CheckoutSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_checkout.CheckoutSummaryResponse'
              type: object
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Get checkout summary
      tags:
      - Checkout
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  SessionAuth:
    in: header
    name: X-Session-ID
    type: apiKey
swagger: "2.0"